new Primus memset implementation
The current memset works quite straightforward and for every byte in the
requested memory it:
1) casts a value to the size of a byte
2) makes an observation before storing
3) stores the byte
4) makes an observation after storing.

The problem is that in the primus-promiscuous mode we often can't aford
this number of primitive operations for every single byte as the length
could be a random value, and a primus machine reaches its limits very
soon that makes any block after such memset unreachable.

As a solution, we add a new function (namely `fill`) into the primus
interpreter, that will interact with the underlying memory without
intermediate interpreter's calls, meaning that the number of expressions
we spend on memset is equal to zero.

A couple of observations is added as well, like usually: filling and filled