fixes taint garbage collector
The `finished` observation may not reach a receiver, because of
unspecified order in which primus components are executed. E.g. greedy
scheduler can intercept `finished` observation and kill a machine. As a
result, no taint is ever finalized, which hits any analysis that depends
on it. And this PR introduces taint finalization on the `halt`
observation, which happens before `finished`.