rectifies x86 primus loader
The problem arises when we deal with libraries: even if a callee is
in the same library, a call will be done through a plt entry first, meaning that
we won't visit the target subroutine.

Our solution is based on the function name: we would better to call the
function itself, but not the corresponding plt entry, so since a
function and a plt entry has the same name, then any call to a plt entry
we substitute to a call to the function.

Also, this PR fixes SP in the `primus-x86-loader` (again).

The history.

PR #993 introduced a proper initialization of PLT entries, but
unfortunately, introduced a bug as well.

Briefly, we didn't want primus to dive deep into PLT entries because
anyway, we would fail with unresolved call exception, but to that moment
a SP would be changed several times.
The solution was to raise an unresolved call exception as soon as
possible, right after a primus machine entered an entry and restore SP
by `unresolved-call` observation.
Unfortunately, the implementation was wrong: we linked an unresolved
call handler to the address of a jump in a plt entry, so actually we
just call a handler instead, and no actual unresolved call happens, no
observation made, and no steps to restore SP taken.

This PR fixes it and now we just unlink all the addresses of plt
entries, so we get a fair unresolved call exception.