enables the Jane Street code style
Previously, we decided to disable the Jane Street code style checks as a
fast solution, but it turned out that it is easier to adopt their style
than to fight with it. To disable the checks, we used the _tags file,
but since we have a global repository the tags file affected all
subprojects, even those that do not use the ppx-jane preprocessor, thus
breaking packages that do not explicitly depend on ppx-jane.

Also, in #957 we accidentally introduced a dependency on a newer version
of `zartith` by using the `Compare` module. We will now use plaincompare
et alas functions so that the old version of `zarith` could be still used
with BAP.