adds the development mode to configure
Adds a new variable called `development` to the configure script that
is disabled by default. When enabled (with the `--enable-development`
option) it will enable certain options and optimizations that are in
general not suitable for the released version. Right now this option,
when enabled, disables cross-module optimizations. We might later
attach more build time optimizations and tweaks to that option.

The cross-module optimizations, when disabled will significantly
improve the rebuilding time (at no to neglible preformance
degradation). Turning them off won't affect a lot, the initial
build time so there is no need to disable them during the initial
installation.