deletes unnecessary OMakefiles
most of the OMakefiles are unmodified so we can just let oasis to
generate them. So far, we have three OMakefiles that we want to have
commited: the root one, that has some global options, and disasm with
llvm OMakefiles, that include C++ and LLVM options which we just
hardcode for now.