adds omake as an optional compilation backend (#1039)
* fixes library dependencies

The OMake plugin is not propagating the transitive closure of
dependencies, so we need to specify our dependencies explicitly and
not depend on dependencies induced by existing dependencies.

* adds OMakefiles

They describe the system and should be committed.
Unfortunately, with OASIS they mostly contain the boilerplate code,
but some of the actually contains the necessary flags.

* cleans up the main Makefile

* updates gitignore to ignore files produced by OMake

* adds a missing header to disasm

* adds the new build_plugins from the build-faster branch

with some changes, e.g., we can't now look for `_build/plugins`

* updates flags

* implements configure-omake

* don't forget the prefix

* fixes the dependency name

* updates missing dependencies

* deletes unnecessary OMakefiles

most of the OMakefiles are unmodified so we can just let oasis to
generate them. So far, we have three OMakefiles that we want to have
commited: the root one, that has some global options, and disasm with
llvm OMakefiles, that include C++ and LLVM options which we just
hardcode for now.

* cleans up the OMakefile

* ignore OMakefiles in general

* runs oasis setup and setup.exe twice

so that files generated by oasis setup could be seen by omake

* cleans up the build_plugins tool

to make codacy happy we switch from using bacticks to `$(...)`.

* builds only directories in the plugins folder

as there could be other files, which are not plugins

* updates .merlin with _buildless paths

since omake puts build artifacts directly into the source tree, we
just need to reference to them without the _build folder.

* fixes penultimated fix

plugins weren't built at all.