switches to meet-only inference with gradual typing
it looks like that having disjoint types is not really useful right
now in Primus Lisp, so we decided to switch to the classical inference
with conjunctive types (i.e., with only meet operator, no join).

It is still the gradual type as we allow the `any` type which relaxes
the typing discipline and allows ill-typed expressions. For example,
the `concat` primitives returns `any` as it is dependetly typed. In
the same vain, we relax the typing of the `if` form and do not unify
the branches, so it returns `any` (so it types depends on the
condition value).

The switch to the simple scheme enables much more readable and useful
error messages, and we can still type all our lisp programs (modulo real
type errors).