restricts the set typing
It was allowed to set an undefined variable, which is probably an
error.

why, probably? The Primus.Env semantics let's us create variables on
the fly. Whether it is used or not, good or not is a second question,
so far let's try to disable this on the typing level and later we will
consider whether we need this in runtime.