prevents cmdliner from fetching plugins path from environment (#1045)
When this option is set, and the `-L` or `--plugin-path` option
is not used, then plugins path will be fetched from the enviroment
variable. However, the plugins library will also look at the same
variable and add the same set of paths to the search. Since this
variable should be local to the plugins library, and in fact is a
part of its interface we should just remove this from the bap-main
command line interface.