adds the  `--primus-lisp-typecheck` option
It will typecheck Primus Lisp programs without the neccesity to run
them (so far we need to run a binary to trigger the typechecker, which
was quite unintuitive).

All error messages will be printed with a conclusive message about the
quality of the program :)

Note, that when a program is run it will be automatically typechecked
and the errors will be printed, so there is no need to use this
option, every time you need to run a program.

Note, for those who prefer to type check, there is also the
`--primus-lisp-type-check` alias. But as usual, the least unambiguous
prefix will work, e.g., `--primus-lisp-type` would suffice.