Revamps primus lisp type checker (#1047)
* rewrites the Gamma and fixes a few bugs

Tasks left to be done:
1) still an open question how to deal with user type variables
2) unresolved variables are not reported
3) unresolved function calls are not reported
4) error message really sucks (at best we have location)

* fixes the scoping of user type variables

A type variable is quantified over the scope of its defining
expression, i.e., for each expression that has a non-ground
signature we instantiate a new type variable of the form `x%m`
where `x` is the name of the type variable and `m` is the id
of the expression.

* adds sane error messages

Well, at least they are more sane than were before :)

1) we track all types that were used during unification and dump them
to users in hope that this will clarify them the issue

2) we disallow unresolved functions and variables and immediately
lower the type to bottom if we see one.

3) the error messages are now more explonatory and include source code
with the relevenat context and highlighting.

* a tweak to underline an offending piece of code on the last line

* adds parameters to the global space

* moves source pretty printing to the source module

also, instead of opening a file, we just keep all the read data in the
source repository.

* exposed the new interface, provided external signatures

* implements diverging type with widen operator

* switches to meet-only inference with gradual typing

it looks like that having disjoint types is not really useful right
now in Primus Lisp, so we decided to switch to the classical inference
with conjunctive types (i.e., with only meet operator, no join).

It is still the gradual type as we allow the `any` type which relaxes
the typing discipline and allows ill-typed expressions. For example,
the `concat` primitives returns `any` as it is dependetly typed. In
the same vain, we relax the typing of the `if` form and do not unify
the branches, so it returns `any` (so it types depends on the
condition value).

The switch to the simple scheme enables much more readable and useful
error messages, and we can still type all our lisp programs (modulo real
type errors).

* ignore definitions that are not applicable to the current context

E.g., we should mips init, when we're checking for x86 (we will get
lots of unbound errors).
```
(defun init (main argc argv auxv)
  "GNU libc initialization stub"
  (declare (external "__libc_start_main")
           (context (arch "mips")))
  (set T9 main)
  (set RA @__libc_csu_fini)
  (exit-with (invoke-subroutine main argc argv)))
```

* restricts the set typing

It was allowed to set an undefined variable, which is probably an
error.

why, probably? The Primus.Env semantics let's us create variables on
the fly. Whether it is used or not, good or not is a second question,
so far let's try to disable this on the typing level and later we will
consider whether we need this in runtime.

* type checks method bodies

(whether a method has a corresponding signal is not yet checked,
comming soon)

* moves signals from the Lisp machine to the Lisp program

* delegates program merging to the program module

* tracks unresolved parameters

* moves typechecker invocation from run to primus-lisp

* some cleanup

* adds signature annotations to all signals.

* moves typechecker into Primus internals

It is now called automatically after components but before init and
type errors are reported as observations. We don't want right now to
stop processing in case of type errors, just in case if we still have
some false positives.

* some clean ups and comments

* delays the plt rewriter in x86-loader until the init

* refactors constant-tracker lisp primitives from the tracker

the primitives shall not be registered conditionally

* drops ids from the printer

* implements C/ABI specific types

They were so far referring to unresolved primitives, now the type
checker says no, so I've reimplemented them not as a primitive,
but as a big overload.

* fixes errors in atoi, discovered by the new typechecker

1) there is no such function as coerce
2) implement atoi using the new language specific size functions.

* adds unless and case macros

and fixes the documentation for `when` and `until`.

* there is no `enter-jmp` signal

* specifies a missing size in `(taint-introduce 'indirectly)`

Previously it was an ill-typed piece of code as we didn't pass the
size. We employ parameters to set the size, and default to one word.

Parameters are dynamically bound, so you can do as

```lisp
(let ((*indirect-taint-size* 100))
   (taint-introduce 'indirectly k v))
```

* fixes a wrongly named variable in `memory/allocate`

* the right name is `fini`

* fixes missing requires and functions

* fixes a missing argument and prunes the undefined-input method

this signal is defined in ddtbd, which is not yet integrated into BAP.

* adds a couple of casts

the `strchr` function takes an integer so we have to explicitly case
chars to it.

* some cleanups and a merlin updates

* adds the  `--primus-lisp-typecheck` option

It will typecheck Primus Lisp programs without the neccesity to run
them (so far we need to run a binary to trigger the typechecker, which
was quite unintuitive).

All error messages will be printed with a conclusive message about the
quality of the program :)

Note, that when a program is run it will be automatically typechecked
and the errors will be printed, so there is no need to use this
option, every time you need to run a program.

Note, for those who prefer to type check, there is also the
`--primus-lisp-type-check` alias. But as usual, the least unambiguous
prefix will work, e.g., `--primus-lisp-type` would suffice.

* updates documentation a little bit