prevents overwriting of a file in the input channel redirection
In Primus Lisp, we have channel redirection which makes files from the
host system visible to the emulation layer. Underneath the hood, we
use common OCaml channels to model files (and probably we should use
Unix file descriptors), and when a file is mapped, we open both input
and output channel. However, when we open the input channel, the file
is truncated, with all the ramifications.

The quickfix is to add the `append` flag. The long-term solution would
be switching to Unix file desciptors and passing flags to the `open(2)`
function.