fixes the SSA algorithm after the Tid graph change
The SSA transformation algorithm assumes that [Sub.to_graph] function
returns a graph with the set of nodes that are equal to the set of
term identifiers of the subroutine. This is no longer true, as this
set is now extended with two artifical nodes, start and exit,
therefore the algorithm was failing.

Now, it expects those artificial nodes and handles them correctly.