tweaks primus-mark-visited to mark called stubs as visited
When a function is stubed by a Primus Lisp function the original plt
entry is not evauluated therefore it is not marked as visited, as a
result we have incorrect statistics on visited blocks.

We now track all primus-lisp-call observations and mark the
corresponding plt entry (or whatever matches with the called address)
as visited.