do not use backend specific registers
Since we don't know the size of a register, we can't create a correct
variable for it (and pulling the register from the depths of the
disassembler would imply too many (possibly breaking) changes), so we
use the register code to represent a register operand instead.

Later, in Primus, we can get the symbolic representation of the
register name using the new `reg-name` primitive. Additionally, we
add the `set-symbol-value` primitive that takes a symbol and assigns
the correspoinding register the passed value. (For common lispers,
`set-symbol-value` is the same as `set` in CL and `setq` is named
`set` in Primus Lisp).

Using these two primitives we can establish a translation between the
disassembler the CPU model and the BAP CPU model in a much more
efficient way.