adds the `lisp-primitive` observation
It is posted every time a primitive is called and returns a value. We
didn't add the usual before primitive observation, mostly because
primitives are not supposed to diverge (and by their nature they
should be implemented using vanilla OCaml and operate on the host CPU
not on the Primus CPU). But we can add one.

Additionally, we didn't provide a signal for this observation as
observing lisp-primitives in lisp will trigger lisp-primitive signals
and is too easy to get an infinite loop in this case.