reifies calls into calls
The IR reification module was translating all jumps to gotos and we
were relying on the later sema lifter pass to translate those gotos
into proper jumps. However, the right thing to do, would be instead
of introducing incorrect term and then fixing it, to produce the
correct reification on the first hand.

Since we have a call method in our Core Theory parse we shall leverage
the knowledge that if [call x] then [is-subroutine x]. Next, in the IR
lifter shall reify to [call x] when we know that [is-subroutine x].
Finally, we have to fix the IR lifter that was relying (incorrectly)
on the behavior that there are no call IR terms returned by our IR
lifter. It was true for the current implementation, but in general
shouldn't be assumed. Several modifications were applied:

1. The `is_call` predicate should take into account intrinsics (in other
words even a `mov` could be a call now) and should be treated that
way. So we added the `has_call` predicate to accompany `is_call`
everywhere (we can join them as they require different contexts)

2. The `turn_into_call` function was assuming that the input is not a
call, now it takes into account that it could be already a call and
respects the call term.

Also, as a bit of refactoring, I've changed `if_conditional` to
`if_unconditional` since we were always using it in the negated
case. (And renamed nonconditional to unconditional as the latter is
more conventional).