provides a new commandline interface for intrinsics
The command line option is now named `--bil-enable-intrinsic` and it
takes a list of instruction classes that we would like to turn into
intrinsic calls. We support, right now, three generic classes, such as
special, unknown, and any. And three parametrized predicates that let
us match with instructions based on the assembly representation and
opcode name or based on their semantics tags.

We decided not to provide a predicate that is based on the binary
representation as it is too low-level and should really go into the
options of corresponding architecture specific plugins, e.g.,
something like `--x86-64-no-vex` is much more appropriate. (Also, we
do not really want to disable all vex, we're pretty fine with vex move
instructions).