adds intrinsics (#1059)
* adds an experimental lifting of unknown instructions

When [--bil-lift-unknown] is set, we will tranlate all instructions
that do not have semantics into calls to intrinsic functions, so that
they later could be stubbed in Primus Lisp.

This is a highly experimental feature, so expect it to change a bit,
before we stabilize the interfaces. It is fully disabled by default,
and only changes something if enabled via the command line inteface
or via the configuration file.

As a working example, let's take
```
40afa1: f3 0f 58 0d 9b dc 00 00                   addss 0xdc9b(%rip), %xmm1
```
which was taken from `/bin/ls` output. When `--bil-lift-unknown`
option is specified, it is lifted as
```
llvm-x86_64:op1 := llvm-x86_64:XMM1
llvm-x86_64:op2 := llvm-x86_64:XMM1
llvm-x86_64:op3 := llvm-x86_64:RIP
llvm-x86_64:op4 := 1
llvm-x86_64:op5 := llvm-x86_64:Nil
llvm-x86_64:op6 := 0xDC9B
llvm-x86_64:op7 := llvm-x86_64:Nil
llvm-x86_64:next_address := 0x40AFA9
call @llvm-x86_64:ADDSSrm with noreturn
```

As we can see, all operands are passed via variables from the
disassembler package (in our case it is llvm-x86_64), since the
meaning of each operand is specific to the backend. All registers,
referenced by an operand are also taken from the backend package to
prevent the name clashing with corresponding registers from the Core
Theory/BIL which can have different meaning/type, e.g., we don't have
XMM registers in BIL, as they are represented as selections from
corresponding YMMS registers. It is expected, that the target support
package should define corresponding methods that will establish
mapping between disassembler specific variables and CPU variables.

Finally, each intrinsic is invoked as a tail-call, so that it can
decide the control flow on itself (i.e., when an instruction is a
control flow instruction on itself). For that we pass the address
following the instruction via the `next_address` which could be
invoked via the `exec-addr` primitive, e.g.

```lisp
(defun adds-stub ()
  (declare (external llvm-x86_64:ADDSSrm))
  (exec-addr llvm-x86_64:next_address))
```

and now every time we have a call to `addss` our adds-stub is called
instead.

* fixes the case macro

the macro-rewriting is not hygienic, keep forgeting that.

* do not use backend specific registers

Since we don't know the size of a register, we can't create a correct
variable for it (and pulling the register from the depths of the
disassembler would imply too many (possibly breaking) changes), so we
use the register code to represent a register operand instead.

Later, in Primus, we can get the symbolic representation of the
register name using the new `reg-name` primitive. Additionally, we
add the `set-symbol-value` primitive that takes a symbol and assigns
the correspoinding register the passed value. (For common lispers,
`set-symbol-value` is the same as `set` in CL and `setq` is named
`set` in Primus Lisp).

Using these two primitives we can establish a translation between the
disassembler the CPU model and the BAP CPU model in a much more
efficient way.

* adds ieee754 support to Primus Lisp

As a demonstration of our new capabilities, we added support for
ieee754 in the form of Primus Lisp primitives and implemented stubs
for three instructions, addsd/rm, divsd/rr, and ucomisd/rm,  each
instruction is about three lines of code, and the full lifter with the
glue code takes 100 lines of code (most of the register mappings).

* adds the `lisp-primitive` observation

It is posted every time a primitive is called and returns a value. We
didn't add the usual before primitive observation, mostly because
primitives are not supposed to diverge (and by their nature they
should be implemented using vanilla OCaml and operate on the host CPU
not on the Primus CPU). But we can add one.

Additionally, we didn't provide a signal for this observation as
observing lisp-primitives in lisp will trigger lisp-primitive signals
and is too easy to get an infinite loop in this case.

* reifies calls into calls

The IR reification module was translating all jumps to gotos and we
were relying on the later sema lifter pass to translate those gotos
into proper jumps. However, the right thing to do, would be instead
of introducing incorrect term and then fixing it, to produce the
correct reification on the first hand.

Since we have a call method in our Core Theory parse we shall leverage
the knowledge that if [call x] then [is-subroutine x]. Next, in the IR
lifter shall reify to [call x] when we know that [is-subroutine x].
Finally, we have to fix the IR lifter that was relying (incorrectly)
on the behavior that there are no call IR terms returned by our IR
lifter. It was true for the current implementation, but in general
shouldn't be assumed. Several modifications were applied:

1. The `is_call` predicate should take into account intrinsics (in other
words even a `mov` could be a call now) and should be treated that
way. So we added the `has_call` predicate to accompany `is_call`
everywhere (we can join them as they require different contexts)

2. The `turn_into_call` function was assuming that the input is not a
call, now it takes into account that it could be already a call and
respects the call term.

Also, as a bit of refactoring, I've changed `if_conditional` to
`if_unconditional` since we were always using it in the negated
case. (And renamed nonconditional to unconditional as the latter is
more conventional).

* relaxes primus type checker wrt to the referenced variables

Now it is possible to reference any variable that is used or defined
in the binary program no matter of the place.

* adds floating point tests

* provides a new commandline interface for intrinsics

The command line option is now named `--bil-enable-intrinsic` and it
takes a list of instruction classes that we would like to turn into
intrinsic calls. We support, right now, three generic classes, such as
special, unknown, and any. And three parametrized predicates that let
us match with instructions based on the assembly representation and
opcode name or based on their semantics tags.

We decided not to provide a predicate that is based on the binary
representation as it is too low-level and should really go into the
options of corresponding architecture specific plugins, e.g.,
something like `--x86-64-no-vex` is much more appropriate. (Also, we
do not really want to disable all vex, we're pretty fine with vex move
instructions).

* uses the new interface for the lisp-primitive observation

* adds symbols to the floats test