adds a stub for realloc (#1035)
* few primus-lisp updates

We add a stub for realloc - surprisingly many allocations are done by
passing a null pointer to realloc, and an analysis that is based on a
tracing of allocation/freeing memory won't notice such call.

Also, this PR fixes a minor bug in the method name in: there is no
signal enter-jmp, so we can't depend on it.

Also, we add a new behavior of the `free` call processing in
memcheck-malloc:  now, if the pointer being freed in unknown, i.e. there
wasn't any memory allocation for it in the current primus machine, then
we allocate one-byte length region and then release it. The
justification is simple: by doing so, we are able to find more bugs like
use-after-free or double-free, e.g. in some library, where any path we
take won't necessarily contain the chain of calls like
`malloc->free->memory-access/free` because such path exists only in the
user code but not in the library itself.

* encoded length

* removed unused function

* rewrote realloc

* refactored a little

* renaming

* removed unresolved names handling from promiscuous