moves taint finalization to the machine-kill observation
All finalization procedures should be invoked in the restricted
mode (i.e., use either machine-kill or system-stop) observations,
as otherwise they risk to have some unpleasant interactions with
other components, e.g., with bap:limit that will kill the machine
when it does any new operation (which increases the machine clock).

In this case it was an infinite loop, as the taint finalization was
attached to the halting observation and was reflected to Primus
Signals which had methods that were doing quite a few
operations, which, in turn, woke up the limited that was using the
halt operation to halt them, and here we go again.